+++ src/Utils/Enums/index.ts
@@ -1,0 +1,32 @@
+export const POSTGRES_ERROR_CODES = {
+  TABLE_NOT_FOUND: '42P01',
+  COLUMN_NOT_FOUND: '42703',
+};
+
+export const MYSQL_ERROR_CODES = {
+  COLUMN_NOT_FOUND: 1054,
+};
+
+export const HTTP_ERROR_CODES = {
+  BAD_REQUEST: 400,
+  UNAUTHORIZED: 401,
+  FORBIDDEN: 403,
+  NOT_FOUND: 404,
+  SERVER_ERROR: 500,
+};
+
+export const ERRORS = {
+  DATABASE: {
+    UNKNOWN: 'E02000',
+    TABLE_NOT_FOUND: 'E02001',
+    COLUMN_NOT_FOUND: 'E02002',
+  },
+  HTTP: {
+    UNKNOWN: 'E01000',
+    NOT_FOUND: 'E01001',
+    UNAUTHORIZED: 'E01002',
+    FORBIDDEN: 'E01003',
+    BAD_REQUEST: 'E01004',
+    SERVER_ERROR: 'E01005',
+  },
+};
